add class for multilabel splitter
