+++ chebai/preprocessing/splitters/__init__.py
@@ -1,3 +1,4 @@
 from .group import GroupSplitter
+from .multilabel import MultiLabelSplitter
 
-__all__ = ["GroupSplitter"]
+__all__ = ["GroupSplitter", "MultiLabelSplitter"]
